Extract USB hand-off flow into usb-handoff.ts to respect the file size cap
